+++ ndlib/ProfileThresholdModel.py
@@ -18,6 +18,10 @@ class ProfileThresholdModel(DiffusionModel):
 
         actual_status = {node: nstatus for node, nstatus in self.status.iteritems()}
 
+        if self.actual_iteration == 0:
+            self.actual_iteration += 1
+            return 0, actual_status
+
         for u in self.graph.nodes():
             if actual_status[u] == 1:
                 continue