+++ ndlib/SISModel.py
@@ -17,6 +17,10 @@ class SISModel(DiffusionModel):
         """
         actual_status = {node: nstatus for node, nstatus in self.status.iteritems()}
 
+        if self.actual_iteration == 0:
+            self.actual_iteration += 1
+            return 0, actual_status
+
         for u in self.graph.nodes():
 
             u_status = self.status[u]
@@ -45,7 +49,7 @@ class SISModel(DiffusionModel):
 
         previous_status = {}
 
-        confidence = 2
+        confidence = 5
         count = 0
 
         while count < max_iteration: